subscription work, code refactor
we now see messages in logs, time to connect them with something to reply with
