+++ pkgs/commands/bot.go
@@ -1,0 +1,3 @@
+package commands
+
+// Reply when certain messages are sent
