+++ pkgs/commands/follows.go
@@ -1,0 +1,3 @@
+package commands
+
+// say something on new follower
