+++ pkgs/commands/sub.go
@@ -1,0 +1,3 @@
+package commands
+
+// say something when new sub
